test: ☑️ refactor route tests
