+++ src/routes/mod.rs
@@ -63,101 +63,3 @@ pub(crate) async fn graphql_handler(
         )
         .into()
 }
-
-#[cfg(test)]
-mod tests {
-    use axum::{
-        body::Body,
-        http::{header, Method, Request, StatusCode},
-        Router,
-    };
-    use http_body_util::BodyExt;
-    use serde_json::{json, Value};
-    use tower::util::ServiceExt;
-
-    use crate::startup::main_router;
-
-    async fn get_app() -> Router {
-        let database_url = "sqlite://:memory:";
-
-        main_router(database_url).await
-    }
-
-    #[tokio::test]
-    async fn graphql_endpoint_returns_200_ok() {
-        // arrange
-        let app = get_app().await;
-
-        // act
-        let response = app
-            .oneshot(Request::builder().uri("/").body(Body::empty()).unwrap())
-            .await
-            .unwrap();
-
-        // assert
-        assert_eq!(response.status(), StatusCode::OK);
-    }
-
-    #[tokio::test]
-    async fn graphql_endpoint_responds_to_invalid_query() {
-        // arrange
-        let app = get_app().await;
-        let json_request_body: Value = json!(
-        {"operationName":"HelloQuery","variables":{},"query":"query HelloQuery { hello "
-            });
-
-        // act
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method(Method::POST)
-                    .uri("/")
-                    .header(header::CONTENT_TYPE, mime::APPLICATION_JSON.as_ref())
-                    .body(Body::from(json_request_body.to_string()))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        // assert
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let body = response.into_body().collect().await.unwrap().to_bytes();
-        let body: Value = serde_json::from_slice(&body).unwrap();
-        assert_eq!(
-            body,
-            json!({
-                "data": None::<String>,
-                "errors": [{
-                    "locations": [{"column": 26,"line": 1}],
-                    "message": " --> 1:26\n  |\n1 | query HelloQuery { hello \n  |                          ^---\n  |\n  = expected selection_set, selection, directive, or arguments"
-                }],
-                "extensions": {"traceId": "00000000000000000000000000000000"}
-            })
-        );
-    }
-
-    #[tokio::test]
-    async fn health_check_returns_expected_json_response_with_200_ok() {
-        // arrange
-        let app = get_app().await;
-
-        // act
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .uri("/health")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        // assert
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let body = response.into_body().collect().await.unwrap().to_bytes();
-        let body: Value = serde_json::from_slice(&body).unwrap();
-        assert_eq!(body, json!({ "healthy": true }));
-    }
-}
