+++ tests/api/main.rs
@@ -3,4 +3,5 @@
 mod helpers;
 mod model;
 mod observability;
+mod routes;
 mod startup;
